Extract out public interface contract to make it easier to update
