+++ server/@types/AvailableComponent.ts
@@ -1,6 +1,6 @@
-export const AVAILABLE_COMPONENTS = ['header', 'footer'] as const
+import type { AvailableComponent } from '../interfaces/externalContract'
 
-export type AvailableComponent = (typeof AVAILABLE_COMPONENTS)[number]
+export const AVAILABLE_COMPONENTS = ['header', 'footer'] as const satisfies AvailableComponent[]
 
 export const isComponent = (component: unknown): component is AvailableComponent =>
   AVAILABLE_COMPONENTS.includes(component as AvailableComponent)
