--- server/@types/Component.ts
@@ -1,5 +1,0 @@
-export default interface Component {
-  html: string
-  css: string[]
-  javascript: string[]
-}
