--- server/interfaces/Service.ts
@@ -1,7 +1,0 @@
-export interface Service {
-  id: string
-  heading: string
-  description: string
-  href: string
-  navEnabled: boolean
-}
