+++ server/routes/developRoutes.ts
@@ -1,9 +1,10 @@
 import { Router } from 'express'
 import { Services } from '../services'
 import authorisationMiddleware from '../middleware/authorisationMiddleware'
-import { AVAILABLE_COMPONENTS, type AvailableComponent } from '../@types/AvailableComponent'
+import { AVAILABLE_COMPONENTS } from '../@types/AvailableComponent'
 import auth from '../authentication/auth'
 import tokenVerifier from '../data/tokenVerification'
+import type { AvailableComponent } from '../interfaces/externalContract'
 import ComponentsController from '../controllers/componentsController'
 import populateCurrentUser from '../middleware/populateCurrentUser'
 import { ComponentRenderer } from '../services/componentRenderer'
