+++ server/services/userService.ts
@@ -3,9 +3,9 @@ import getServicesForUser from './utils/getServicesForUser'
 import CacheService from './cacheService'
 import { ServiceActiveAgencies } from '../@types/activeAgencies'
 import config from '../config'
-import { PrisonUser, PrisonUserAccess } from '../interfaces/hmppsUser'
-import { Service } from '../interfaces/Service'
-import { PrisonCaseload } from '../interfaces/caseLoad'
+import type { PrisonCaseload } from '../interfaces/caseLoad'
+import type { Service } from '../interfaces/externalContract'
+import type { PrisonUser, PrisonUserAccess } from '../interfaces/hmppsUser'
 import AllocationsApiClient, { StaffAllocationPolicies } from '../data/AllocationsApiClient'
 import { Role } from './utils/roles'
 import LocationsInsidePrisonApiClient from '../data/locationsInsidePrisonApiClient'
